test: QuestionCommonUpdateController Test 추가
- 기존 공통된 Question Update관련 로직 테스트를 구현
