--- application/src/main/java/com/comssa/api/question/license/service/AdminLicenseMuiltipleChoiceQuestionUpdateService.java
@@ -1,26 +1,0 @@
-package com.comssa.api.question.license.service;
-
-import com.comssa.api.question.common.service.QuestionUpdateService;
-import com.comssa.persistence.question.license.domain.LicenseMultipleChoiceQuestion;
-import com.comssa.persistence.question.license.repository.LicenseMultipleChoiceQuestionRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-@Transactional
-@RequiredArgsConstructor
-public class AdminLicenseMuiltipleChoiceQuestionUpdateService
-	implements QuestionUpdateService<LicenseMultipleChoiceQuestion> {
-	private final LicenseMultipleChoiceQuestionRepository licenseMultipleChoiceQuestionRepository;
-
-	@Override
-	public void deleteQuestion(Long questionId) {
-		licenseMultipleChoiceQuestionRepository.deleteById(questionId);
-	}
-
-	@Override
-	public LicenseMultipleChoiceQuestion findById(Long questionId) {
-		return licenseMultipleChoiceQuestionRepository.findByIdFetchChoices(questionId).orElse(null);
-	}
-}
