+++ application/src/main/resources/static/js/license-index/updateQuestion.js
@@ -30,7 +30,7 @@ function toggleCategory(categoryName) {
 
 function deleteQuestion(questionId) {
     if (confirm("정말로 이 질문을 삭제하시겠습니까?")) {
-        fetch(`/admin/question/license/${questionId}`, {
+        fetch(`/admin/question/common/${questionId}`, {
             method: 'DELETE',
         })
             .then(response => {
